Work on edit courses and edit profile
