--- adminAccountManagment.py
@@ -1,32 +1,0 @@
-from django.contrib.auth.models import User
-from django.core.exceptions import ObjectDoesNotExist
-
-from scheduler.models import UserTable, LabTable, CourseTable
-
-
-class AdminAccountManagementPage:
-    def __init__(self):
-        pass
-
-    @staticmethod
-    def deleteAccount(usernameID, emailID):
-        try:
-            # Get the user account based on username and email
-            account = User.objects.get(id=usernameID)
-            user = UserTable.objects.filter(email=account.email).first()
-            if (account.email != user.email) | (usernameID != emailID):
-                return "username/email match error"
-
-            # Delete children
-            if user.userType == "ta":
-                LabTable.objects.filter(taId=user.id).delete()
-            elif user.userType == "instructor":
-                CourseTable.objects.filter(instructorId=user.id).delete()
-
-            # Finally delete user...
-            user.delete()
-            # Delete the account
-            account.delete()
-            return "Account deleted successfully"
-        except ObjectDoesNotExist:
-            return "Failed to delete account"
