--- static/css/createAccount.css
@@ -1,74 +1,0 @@
-.body {
-    font-family: Arial, sans-serif;
-    /*Remove if needed1*/
-    background-image: url('static/images/university-of-wisconsin-milwaukee.jpg');
-    background-color: #f7f7f7;
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    /*Remove if needed2*/
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.createAccount-container {
-    text-align: center;
-    background: #fff;
-    padding: 40px;
-    margin: auto;
-    width: 300px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.createAccount-container h1 {
-    font-size: 24px;
-    margin-bottom: 20px;
-}
-
-.createAccount-container form {
-    margin-bottom: 15px;
-}
-
-.createAccount-container input[type="text"],
-.createAccount-container input[type="password"],
-.createAccount-container input[type="email"] {
-    width: 100%;
-    padding: 10px;
-    margin: 10px 0;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    box-sizing: border-box;
-}
-
-.createAccount-container input[type="submit"] {
-    width: 100%;
-    padding: 10px;
-    margin-top: 10px;
-    border: none;
-    border-radius: 4px;
-    background-color: #333;
-    color: white;
-    cursor: pointer;
-}
-
-.createAccount-container input[type="submit"]:hover {
-    background-color: #555;
-}
-
-.createAccount-container p {
-    margin-top: 20px;
-}
-
-/* Adjust the color of links */
-.createAccount-container a {
-    color: #0067ab;
-    text-decoration: none;
-}
-
-.createAccount-container a:hover {
-    text-decoration: underline;
-}
